+++ bin/faucet/src/config.rs
@@ -33,13 +33,16 @@ pub struct FaucetConfig {
     pub faucet_account_path: PathBuf,
     /// The salt to be used by the server to generate the `PoW` seed
     pub pow_salt: String,
+    /// Optional: Endpoint of the remote transaction prover in the format
+    /// `<protocol>://<host>[:<port>]`
+    pub remote_tx_prover_url: Option<Url>,
 }
 
 impl Display for FaucetConfig {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         f.write_fmt(format_args!(
-            "{{ endpoint: \"{}\", node_url: \"{}\", timeout_ms: \"{}\", asset_amount_options: {:?}, faucet_account_path: \"{}\", pow_salt: \"{}\" }}",
-            self.endpoint, self.node_url, self.timeout_ms, self.asset_amount_options, self.faucet_account_path.display(), self.pow_salt
+            "{{ endpoint: \"{}\", node_url: \"{}\", timeout_ms: \"{}\", asset_amount_options: {:?}, faucet_account_path: \"{}\", pow_salt: \"{}\", remote_tx_prover_url: \"{:?}\" }}",
+            self.endpoint, self.node_url, self.timeout_ms, self.asset_amount_options, self.faucet_account_path.display(), self.pow_salt, self.remote_tx_prover_url
         ))
     }
 }
@@ -56,6 +59,7 @@ impl Default for FaucetConfig {
             asset_amount_options: AssetOptions::new(vec![100, 500, 1_000]).unwrap(),
             faucet_account_path: DEFAULT_FAUCET_ACCOUNT_PATH.into(),
             pow_salt: "miden-faucet".to_string(),
+            remote_tx_prover_url: None,
         }
     }
 }