+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Added a required Proof-of-Work in the faucet to request tokens (#831).
+- Faucet now supports the usage of a remote transaction prover (#830).
 
 ## v0.8.2 (2025-05-04)